refactor: standardize error messages

+++ src/modules/videos/useCases/DeleteVideo/DeleteVideoUseCase.ts
@@ -1,6 +1,7 @@
 import { injectable, inject } from 'tsyringe';
 
 import { IVideosRepository } from '@modules/videos/repositories/IVideosRepository';
+import { AppError } from '@shared/errors/AppError';
 
 @injectable()
 class DeleteVideoUseCase {
@@ -10,6 +11,12 @@ class DeleteVideoUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    const videoExists = await this.videosRepository.findVideoById(id);
+
+    if (!videoExists) {
+      throw new AppError('video not found', 404);
+    }
+
     await this.videosRepository.deleteVideo(id);
   }
 }
